requirements.txt
Django==2.2
django-filter==2.2.0
djangorestframework==3.10.3
djangorestframework-simplejwt==4.3.0
mysqlclient==1.4.4
Pillow==6.2.1
pkg-resources==0.0.0
PyJWT==1.7.1
pytz==2019.3
sqlparse==0.3.0
